Minor tweaks for dynamic_format_arg_store
